Workaround for travis boto issue with GCE
"Yeah so this happens because of the /etc/boto.cfg file which (I believe)
Google adds to their system images - the idea being that it loads some plugins
that let you use boto more easily with Google Cloud services. This is mostly
useless in a CI environment though. If you look at the config file, you'll see
that it tries to load a Python plugin, which the travis people must have
somehow corrupted in this image. It is unfortunate that they switched to a new
image without testing widely used libraries like boto."

https://github.com/travis-ci/travis-ci/issues/7940